re-organize join_bias code into self-contained module. reduce repetition
